changes and improves create sale page
Some style improvements. Several selects are gated on condition of other
selects having user selected values. Commissions now properly calculate
whether a discount is applicable depending on the sale date selected,
and update in real-time on a date change while a salesperson and product
are selected. The form is reset on a create sale (except date).